line-bot-test/requirements.txt
Flask==2.3.2
beautifulsoup4==4.12.2
requests==2.28.2
line-bot-sdk==2.4.2